add tests for delete location functionality -refactor locationRoute test
